Fix warning when comparing qsizetype to result of sizeof
   warning: comparison of integers of different signs: 'qsizetype' (aka 'long long') and 'unsigned long'

+++ tests/auto/corelib/plugin/qpluginloader/tst_qpluginloader.cpp
@@ -384,7 +384,7 @@ void tst_QPluginLoader::loadMachO()
         return;
 
     QVERIFY(pos > 0);
-    QVERIFY(len >= sizeof(void*));
+    QVERIFY(size_t(len) >= sizeof(void*));
     QVERIFY(pos + long(len) < data.size());
     QCOMPARE(pos & (sizeof(void*) - 1), 0UL);
